requirements.txt
aiohttp==3.6.2
apple-music-python==1.0.3
applescript==2020.12.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cryptography==3.3.1
cycler==0.11.0
discord-ui==5.1.6
discord.py==2.0.0a3575+g45d498c
Flask==1.1.2
fonttools==4.30.0
gitdb==4.0.5
GitPython==3.1.11
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
itunes==2020.12.3
Jinja2==2.11.2
kiwisolver==1.3.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.5.1
mccabe==0.6.1
mcipc==2.3.8
multidict==4.7.6
numpy==1.22.3
packaging==21.3
Pillow==9.0.1
ping3==2.6.6
py-cord==2.0.0a4631+g16f9bcb5
pycparser==2.20
PyJWT==1.7.1
pylint==2.6.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.20.0
rcon==2.1.1
requests==2.25.1
six==1.15.0
smmap==3.0.4
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.2
utilsx==0.3.53
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.0
